Fixed flaky test: WebRtcVideoChannel2BaseTest.GetStatsMultipleRecvStreams
This test has been failing on TSan lately:
  ../../webrtc/media/base/videoengine_unittest.h:519: Failure
  Value of: GetReceiverStats(i).frame_width
    Actual: 0
    Expected: DefaultCodec().width
    Which is: 640

The root cause for the failure appears to be that the stats update
(https://cs.chromium.org/chromium/src/third_party/webrtc/video/video_receive_stream.cc?rcl=1472584967&l=353)
happens to be after the frame is passed to the renderer - while the test
is only waiting for the former.

The fix is to give it some extra time using EXPECT_EQ_WAIT instead.

Review-Url: https://codereview.webrtc.org/2299483002
Cr-Commit-Position: refs/heads/master@{#13991}