Reduce full stack test time to 45 secs and add H264 and FlexFEC.
This CL adds full stack tests that are used to measure the performance
of H264 with and without FlexFEC. In order to not increase the bot
run time, the CL also reduces the test time to 45 secs. This should
be OK, since the BWE is faster to ramp up nowadays.

Due to the test time change, there may be some performance alerts.

BUG=webrtc:5654

Review-Url: https://codereview.webrtc.org/2499273002
Cr-Commit-Position: refs/heads/master@{#15118}