Optionally disable APM limiter in AudioMixer.
The APM limiter is a component for keeping the audio from clipping by smoothly reducing the amplitude of the audio samples. It can be rather expensive because of band-splitting & merging. Also, experiments indicate that it is of questionable benefit (adding several sources of human speech almost never cause clipping).

To optionally disable the limiter, this CL does some refactoring on the (quite large) AudioMixerImpl. Functionality related to actual addition of frames and handling AudioFrame meta-data (sample_rate, num_channels, samples_per_channel, time_stamp, elapsed_time_ms) is broken out in a new sub-component called FrameCombiner.

The FrameCombiner is initialized with a 'use_limiter' flag. To create a mixer without using the APM limiter

Inside of FrameCombiner, the meta-data handling and the audio sample addition are kept divided from each other.

This also fixes a few minor GN issues so that warnings do not have to be suppressed.

BUG=webrtc:7167

Review-Url: https://codereview.webrtc.org/2692333002
Cr-Commit-Position: refs/heads/master@{#16742}