Normalize codec names to those used by AcmCodecDatabase.
With the upcoming changes to WebRtcVoiceEngine due to injectable audio
encoders, codec names will no longer be normalized against the ACM's
codec database. This not only breaks a couple of downstream tests
(which I've written fixes for) but also runs the risk of breaking
other external usage, like SDP mangling etc.

Eventually, we should change to using the spelling from the relevant
RFCs, although not as part of this other, large change.

BUG=webrtc:5806

Review-Url: https://codereview.webrtc.org/2845603002
Cr-Commit-Position: refs/heads/master@{#17903}