Removed implicit divisions in the residual echo detector
This CL changes the updating of a loop index from using the modulus
operator to using a conditional, avoiding the divisions done in the
modulus operator which had a measurable impact on mpbile
platforms.

Additionally, there is a similar removal of another modulus operator, but the impact of that should be negligible.

BUG=webrtc:7666

Review-Url: https://codereview.webrtc.org/2882183004
Cr-Commit-Position: refs/heads/master@{#18168}