Try creating sockets again if network change occurs after bind failed.
If the network interface appears active, but binding the sockets fails,
then it won't produce any candidates even though it's never marked as
"network failed". So this was causing nothing to happen once a network
change event occurs and the interface becomes usable again.

So, this CL adds the condition that we only disable gathering of local
ports if we don't have them already.

See bug for more details.

BUG=webrtc:8256

Review-Url: https://codereview.webrtc.org/3015543002
Cr-Commit-Position: refs/heads/master@{#20007}