Make PictureIdTest more strict.
Only allow gaps in picture id for key frames.
When a VideoSendStream is destroyed, frames in the queue not yet sent are lost. The recreated stream
should start with a key frame.

Also enable PictureIdIncreasingAfterStreamCountChangeSimulcastEncoderAdapter if forced fallback is
enabled. In this case, the picture id is set in the PayloadRouter and the sequence should be
continuous.


Bug: none
Change-Id: If7987166c86d6a8edbe5e479701f7f04c49cd89c
Reviewed-on: https://webrtc-review.googlesource.com/7363
Commit-Queue: Åsa Persson <asapersson@webrtc.org>
Reviewed-by: Rasmus Brandt <brandtr@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#20216}