Remove an unnecessary non-const-iterator function.
With an old C++-11 std-library, this gave compilation errors. A const
iterator couldn't be used in a non-const setting. One couldn't call
vector.erase(iter) on it. Now it seems we have updated to a less old
C++ standard lib version (it passes on the bots).

Bug: None
Change-Id: Iae37d49ed06e1635f55c8a31651171585aea6d8d
Reviewed-on: https://webrtc-review.googlesource.com/36160
Reviewed-by: Sam Zackrisson <saza@webrtc.org>
Commit-Queue: Alex Loiko <aleloi@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#21431}