Change levels of different speech signal in tool.
The conversational_speech_generator tool now adjusts the level of
different speech segments.

Implementation:
The Turn and MultiEndCall::SpeakingTurn structs have an extra 'gain'
member.  It's read and parsed in timing.cc and put in a Turn
struct. It's put in a SpeakingTurn struct in multiend_call.cc and read
and applied to the signal in simulator.cc

Bug: webrtc:7494
Change-Id: I9b82a896eb616c8b5ef14d41dfdfd085ef1d3fbb
Reviewed-on: https://webrtc-review.googlesource.com/26280
Commit-Queue: Alex Loiko <aleloi@webrtc.org>
Reviewed-by: Alessio Bazzica <alessiob@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#21714}