Enable SVC based on number of SSRCs.
Number of spatial layers is set equal to number of SSRCs. The maximum
value is limited to 3. If spatial layering is enabled, i.e. number of
spatial layers is greater than 1, then number of temporal layers is set
to 3. Otherwise number of temporal layers is set to 1.

Number of spatial and temporal layers can be overwritten through field
trial.

Bug: webrtc:8931
Change-Id: I37bd7fe053529683dc3e91b4e544fbdb44429340
Reviewed-on: https://webrtc-review.googlesource.com/59440
Commit-Queue: Sergey Silkin <ssilkin@webrtc.org>
Reviewed-by: Stefan Holmer <stefan@webrtc.org>
Reviewed-by: Erik Språng <sprang@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#22511}