Adds field trial for disabling pacer queue draining.
This CL adds a field trial that disables the feature that the pacer will
ignore the pacing rate and send extra fast to drain the queues if the
pacer queue starts to fill up. BBR assumes that the pacing rate will be
respected and sending more increase the risk of overestimating the
bandwidth.

Bug: webrtc:8415
Change-Id: Ibba315360dafef1c317d14a83199172f9f8cc6aa
Reviewed-on: https://webrtc-review.googlesource.com/80964
Reviewed-by: Stefan Holmer <stefan@webrtc.org>
Reviewed-by: Björn Terelius <terelius@webrtc.org>
Commit-Queue: Sebastian Jansson <srte@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#23773}