Implement changing degradation preference with setParameters()
The current default behavior is unchanged and points to MAINTAIN_FRAMERATE,
meaning there is no way to currently use BALANCED as we can't detect
when the value as been set or not.
Updating this is an API change that should be done in another CL and
properly communicated first.


Bug: webrtc:7607
Change-Id: Ic3877ad8dd7bc418296f21a04bc37f59ec55934a
Reviewed-on: https://webrtc-review.googlesource.com/88766
Reviewed-by: Stefan Holmer <stefan@webrtc.org>
Reviewed-by: Steve Anton <steveanton@webrtc.org>
Commit-Queue: Florent Castelli <orphis@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#24024}